Move MetricalDurationSpanningContainerBuilder to own file
